+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/JibExtension.java
@@ -64,9 +64,11 @@ import org.gradle.api.tasks.Optional;
  *   }
  *   allowInsecureRegistries = false
  *   containerizingMode = 'exploded'
- *   jibExtensions {
- *     implementation = 'com.example.ThirdPartyJibGradleExtension'
- *     properties = [customKey: 'value]
+ *   pluginExtensions {
+ *     pluginExtension {
+ *       implementation = 'com.example.ThirdPartyJibGradleExtension'
+ *       properties = [customKey: 'value]
+ *     }
  *   }
  * }
  * }</pre>
@@ -77,8 +79,6 @@ public class JibExtension {
   private static final boolean DEFAULT_ALLOW_INSECURE_REGISTIRIES = false;
   private static final String DEFAULT_CONTAINERIZING_MODE = "exploded";
 
-  private final Project project;
-
   private final BaseImageParameters from;
   private final TargetImageParameters to;
   private final ContainerParameters container;
@@ -88,7 +88,8 @@ public class JibExtension {
   private final SkaffoldParameters skaffold;
   private final Property<Boolean> allowInsecureRegistries;
   private final Property<String> containerizingMode;
-  private final ListProperty<ExtensionsParameters> jibExtensions;
+  private final ListProperty<ExtensionParameters> pluginExtensions;
+  private final ExtensionParametersSpec extensionParametersSpec;
 
   /**
    * Should be called using {@link org.gradle.api.plugins.ExtensionContainer#create}.
@@ -96,7 +97,6 @@ public class JibExtension {
    * @param project the injected gradle project
    */
   public JibExtension(Project project) {
-    this.project = project;
     ObjectFactory objectFactory = project.getObjects();
 
     from = objectFactory.newInstance(BaseImageParameters.class);
@@ -107,7 +107,9 @@ public class JibExtension {
     outputPaths = objectFactory.newInstance(OutputPathsParameters.class, project);
     skaffold = objectFactory.newInstance(SkaffoldParameters.class, project);
 
-    jibExtensions = objectFactory.listProperty(ExtensionsParameters.class).empty();
+    pluginExtensions = objectFactory.listProperty(ExtensionParameters.class).empty();
+    extensionParametersSpec =
+        objectFactory.newInstance(ExtensionParametersSpec.class, project, pluginExtensions);
     allowInsecureRegistries = objectFactory.property(Boolean.class);
     containerizingMode = objectFactory.property(String.class);
 
@@ -144,15 +146,8 @@ public class JibExtension {
     action.execute(skaffold);
   }
 
-  /**
-   * Adds a new extension configuration to the extensions list.
-   *
-   * @param action closure representing an extension configuration
-   */
-  public void jibExtensions(Action<? super ExtensionsParameters> action) {
-    ExtensionsParameters extension = project.getObjects().newInstance(ExtensionsParameters.class);
-    action.execute(extension);
-    jibExtensions.add(extension);
+  public void pluginExtensions(Action<? super ExtensionParametersSpec> action) {
+    action.execute(extensionParametersSpec);
   }
 
   public void setAllowInsecureRegistries(boolean allowInsecureRegistries) {
@@ -222,7 +217,7 @@ public class JibExtension {
 
   @Nested
   @Optional
-  public ListProperty<ExtensionsParameters> getJibExtensions() {
-    return jibExtensions;
+  public ListProperty<ExtensionParameters> getPluginExtensions() {
+    return pluginExtensions;
   }
 }