+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/JavaContainerBuilderHelper.java
@@ -42,14 +42,16 @@ public class JavaContainerBuilderHelper {
   /**
    * Returns a {@link FileEntriesLayer} for adding the extra directory to the container.
    *
-   * @param extraDirectory the source extra directory path
+   * @param sourceDirectory the source extra directory path
+   * @param targetDirectory the root directory on the container to place the files in
    * @param extraDirectoryPermissions map from path on container to file permissions
    * @param modificationTimeProvider file modification time provider
    * @return a {@link FileEntriesLayer} for adding the extra directory to the container
    * @throws IOException if walking the extra directory fails
    */
   public static FileEntriesLayer extraDirectoryLayerConfiguration(
-      Path extraDirectory,
+      Path sourceDirectory,
+      AbsoluteUnixPath targetDirectory,
       Map<String, FilePermissions> extraDirectoryPermissions,
       BiFunction<Path, AbsoluteUnixPath, Instant> modificationTimeProvider)
       throws IOException {
@@ -61,12 +63,12 @@ public class JavaContainerBuilderHelper {
           FileSystems.getDefault().getPathMatcher("glob:" + entry.getKey()), entry.getValue());
     }
 
-    new DirectoryWalker(extraDirectory)
+    new DirectoryWalker(sourceDirectory)
         .filterRoot()
         .walk(
             localPath -> {
               AbsoluteUnixPath pathOnContainer =
-                  AbsoluteUnixPath.get("/").resolve(extraDirectory.relativize(localPath));
+                  targetDirectory.resolve(sourceDirectory.relativize(localPath));
               Instant modificationTime = modificationTimeProvider.apply(localPath, pathOnContainer);
               FilePermissions permissions =
                   extraDirectoryPermissions.get(pathOnContainer.toString());